@@ -14,21 +14,25 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "axios": "^0.23.0",
-    "cookie-parser": "^1.4.5",
+    "axios": "^1.4.0",
+    "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
-    "dayjs": "^1.8.32",
-    "dotenv": "^8.2.0",
+    "dayjs": "^1.11.8",
+    "dotenv": "^16.1.4",
     "express": "^4.17.1",
-    "jsonwebtoken": "^8.5.1",
-    "mongoose": "^5.9.18",
-    "query-string": "^6.13.1",
-    "simple-statistics": "^7.1.0"
+    "jsonwebtoken": "^9.0.0",
+    "mongoose": "^7.2.4",
+    "nodemon": "^2.0.6",
+    "path": "^0.12.7",
+    "query-string": "^8.1.0",
+    "simple-statistics": "^7.8.3",
+    "url": "^0.11.1"
   },
   "devDependencies": {
     "babel-eslint": "^10.1.0",
-    "chalk": "^4.1.0",
-    "concurrently": "^6.3.0",
-    "eslint": "^7.2.0"
-  }
+    "chalk": "^5.2.0",
+    "concurrently": "^8.2.0",
+    "eslint": "^8.42.0"
+  },
+  "type": "module"
 }
