@@ -1,52 +1,32 @@
-require('dotenv').config()
+import 'dotenv/config'
 
-const env = process.env.NODE_ENV
-
-const development = {
+const config = {
   app: {
-    port: 5000,
+    port: parseInt(process.env.SERVER_PORT),
     secret_key: process.env.SECRET_KEY,
   },
   api: {
     client_id: process.env.CLIENT_ID,
     client_secret: process.env.CLIENT_SECRET,
-    redirect_uri: 'http://localhost:5000/auth/spotify/callback/',
-  },
-  db: {
-    host: 'localhost',
-    port: 27017,
-    name: 'theRestOfTheIceberg',
-  },
-  env: env,
-}
-
-const production = {
-  app: {
-    port: parseInt(process.env.PORT),
-    secret_key: process.env.SECRET_KEY,
-  },
-  api: {
-    client_id: process.env.CLIENT_ID,
-    client_secret: process.env.CLIENT_SECRET,
-    redirect_uri: process.env.REDIRECT_URI,
   },
   db: {
     username: process.env.MONGO_USERNAME,
     password: process.env.MONGO_PASSWORD,
-    host: 'cluster0.3bvad.mongodb.net',
-    port: 27017,
-    name: 'theRestOfTheIceberg',
+    host: process.env.MONGO_HOST,
     options: {
       retryWrites: true,
       w: 'majority',
     },
   },
-  env: env,
+  env: process.env.NODE_ENV,
 }
 
-const config = {
-  development,
-  production,
+if (config.env === 'development') {
+  config.api.redirect_uri = process.env.REDIRECT_URI_DEV + ':' + process.env.SERVER_PORT + '/auth/spotify/callback/'
+} else if (config.env === 'production') {
+  config.api.redirect_uri = process.env.REDIRECT_URI_PROD + ':' + process.env.SERVER_PORT + '/auth/spotify/callback/'
+} else {
+
 }
 
-module.exports = config[env]
+export default config
